separated cards component into multiple files
created separate files for cards components
