+++ app/frontend/src/components/Cards/InnerCopNavCard.js
@@ -5,18 +5,6 @@ import React from "react";
 // Internal Imports
 import { combineClasses } from "../Utility/utils";
 
-function InnerCopCard(props) {
-  return (
-    <div className={combineClasses("card", "inner-cop-card", props.addClass)}>
-      <div className="inner-cop-card-content">{props.children}</div>
-    </div>
-  );
-}
-
-InnerCopCard.propTypes = {
-  addClass: PropTypes.string,
-};
-
 function InnerCopNavCard({ isActive = false, ...props }) {
   const Tag = "href" in props ? "a" : "button"; // conditionally rendered tagss
 
@@ -42,4 +30,4 @@ InnerCopNavCard.propTypes = {
   onClick: PropTypes.func,
 };
 
-export { InnerCopCard, InnerCopNavCard };
+export { InnerCopNavCard };
