+++ app/frontend/src/components/Navigation/InnerCopNav.js
@@ -10,7 +10,7 @@ import {
   CopIconProduct,
   CopIconUiux,
 } from "../../assets/images/images";
-import { InnerCopNavCard } from "../Cards/Cards";
+import { InnerCopNavCard } from "../Cards/InnerCopNavCard";
 import * as utility from "../Utility/utils";
 
 function InnerCopNav(props) {
